# docs/swagger.yaml
basePath: /api
definitions:
...
  userhandler.CreateUserPayload:
    properties:
      is_admin:
        type: boolean
      password:
        type: string
      permissions:
        items:
          $ref: '#/definitions/userhandler.DatabasePermission'
        type: array
      username:
        type: string
    type: object
  userhandler.DatabasePermission:
    properties:
      can_create:
        type: boolean
      can_delete:
        type: boolean
      can_edit:
        type: boolean
      can_view:
        type: boolean
      database_name:
        type: string
    type: object
  userhandler.UpdateMePayload:
    properties:
      new_password:
        type: string
      old_password:
        type: string
    type: object
  userhandler.UpdateUserPayload:
    properties:
      is_admin:
        type: boolean
      password:
        type: string
      permissions:
        items:
          $ref: '#/definitions/userhandler.DatabasePermission'
        type: array
    type: object
  userhandler.UserResponse:
    properties:
      id:
        type: integer
      is_admin:
        type: boolean
      permissions:
        items:
          $ref: '#/definitions/userhandler.DatabasePermission'
        type: array
      username:
        type: string
    type: object
  utils.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  utils.MessageResponse:
    properties:
      message:
        type: string
    type: object
info:
  contact:
    email: denglerchr@gmail.com
    name: Christian Dengler
    url: https://www.swcd.lu
  description: This is a server for a image, audio and file storage with integrated
    web-ui.
  title: SWCD MediaHub-API
  version: 2.0.0-beta1
paths:
...
  /me:
    get:
      description: Retrieves the user record and their specific permissions for the
        currently authenticated user. Admins will receive an empty permissions array.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userhandler.UserResponse'
        "401":
          description: Authentication failed
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Retrieve current user
      tags:
      - User
    patch:
      consumes:
      - application/json
      description: Updates the password for the currently authenticated user. Requires
        the old password to verify identity.
      parameters:
      - description: Old and New Password
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/userhandler.UpdateMePayload'
      produces:
      - application/json
      responses:
        "200":
          description: Password updated successfully
          schema:
            $ref: '#/definitions/utils.MessageResponse'
        "400":
          description: Invalid JSON body or missing fields
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: 'Authentication failed: invalid old password'
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Update current user's password
      tags:
      - User
  /user:
    delete:
      description: Deletes a user account and cascades the deletion to remove their
        database permissions. Requires the global IsAdmin role.
      parameters:
      - description: User ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User successfully deleted
          schema:
            $ref: '#/definitions/utils.MessageResponse'
        "400":
          description: Missing or invalid ID format
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Authentication failed
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: 'Forbidden: Admin user not retrieved'
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "409":
          description: Cannot delete the last remaining admin user
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Delete a user
      tags:
      - User
    patch:
      consumes:
      - application/json
      description: Updates an existing user's global status, password, or database
        permissions. Permissions act as an Upsert/Replace operation. Requires the
        global IsAdmin role.
      parameters:
      - description: User ID
        in: query
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/userhandler.UpdateUserPayload'
      produces:
      - application/json
      responses:
        "200":
          description: User successfully updated
          schema:
            $ref: '#/definitions/userhandler.UserResponse'
        "400":
          description: Missing ID or invalid JSON body
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Authentication failed
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: 'Forbidden: Admin user not retrieved'
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "409":
          description: Cannot remove last admin user
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Update an existing user
      tags:
      - User
    post:
      consumes:
      - application/json
      description: Creates a new user account and optionally defines their database
        permissions. Requires the global IsAdmin role.
      parameters:
      - description: New user details and permissions
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/userhandler.CreateUserPayload'
      produces:
      - application/json
      responses:
        "201":
          description: User successfully created
          schema:
            $ref: '#/definitions/userhandler.UserResponse'
        "400":
          description: Invalid JSON body, missing fields, or duplicate database names
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Authentication failed
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: 'Forbidden: Admin user not retrieved'
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "409":
          description: User already exists
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Create a new user
      tags:
      - User
  /users:
    get:
      description: Retrieves a list of all user accounts and their assigned database
        permissions. Requires the global IsAdmin role.
      produces:
      - application/json
      responses:
        "200":
          description: List of users
          schema:
            items:
              $ref: '#/definitions/userhandler.UserResponse'
            type: array
        "401":
          description: Authentication failed
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: 'Forbidden: User lacks IsAdmin role'
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Retrieve all users
      tags:
      - User
schemes:
- http
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    description: Type "Bearer" followed by a space and a JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"